Only list's admin can see the "add" button The button is visible only in Articles component Retrieve from Firebase whether user is an admin or not

+++ src/app/articles/articles.component.ts
@@ -1,8 +1,9 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnInit, ViewChild} from '@angular/core';
 import {AmenitiesService} from "../amenities.service";
 import {Observable} from "rxjs/Observable";
 import {Article, CAT_TO_IMAGE} from "../article";
 import {ActivatedRoute, Params} from "@angular/router";
+import {EditorComponent} from "../editor/editor.component";
 
 @Component({
   selector: 'app-articles',
@@ -10,8 +11,10 @@ import {ActivatedRoute, Params} from "@angular/router";
   styles: [``]
 })
 export class ArticlesComponent implements OnInit {
+  @ViewChild(EditorComponent) editor: EditorComponent;
 
   public articles: Observable<Article[]>;
+  public isAdmin: boolean;
   private selectedList: string;
 
   constructor(private amenitiesService: AmenitiesService, private route: ActivatedRoute) {
@@ -21,6 +24,7 @@ export class ArticlesComponent implements OnInit {
     this.route.params.subscribe(params => {
       this.selectedList = params['listName'];
       this.articles = this.amenitiesService.getArticles(this.selectedList);
+      this.amenitiesService.isAdmin(this.selectedList).subscribe(admin => this.isAdmin = admin); // TODO replace by switchMap
     });
   }
 
@@ -31,9 +35,19 @@ export class ArticlesComponent implements OnInit {
   }
 
   removeArticle(id: string) {
-    if (this.selectedList) {
+    if (this.selectedList && this.isAdmin) {
       this.amenitiesService.remove(this.selectedList, id);
     }
+  }
+
+  create(): void {
+    if (this.isAdmin) {
+      this.editor.open();
+    }
+  }
+
+  persist(article: Article) {
+    this.amenitiesService.add(article, this.selectedList);
   }
 
   getImage(article: Article): string {
